renaming, folder structure, pattern per field
